+++ CHANGELOG.md
@@ -1,8 +1,14 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
-## [1.16.1] 26th June, 2025
+## [1.17.0] 6th July, 2025
 - Added Delivery Preference Support for Web Popups
+
+## [1.16.2] 3rd July 2025
+- Introduces the new templates 'Advanced Web-popup Builder' in 'Web Popup'.
+
+## [1.16.1] 1st July 2025
+- Removed URL check from Visual Editor rendering
 
 ## [1.16.0] 27th June 2025
 - Added Support for Accessiblity in Soft Prompt